refactor(agent): satisfy source complexity ratchet
